==> weather/step-2-data-prep.ipynb <==
import pandas as pd
import numpy as np
import gc
import swifter
from tqdm.auto import tqdm
import math
#import dask.dataframe as dd

latitude_dtype = np.int8
longitude_dtype = np.int16

### Process Image metadata

- Round the geocode for accuracy level to 111km
- Keep images 2007-2012
- Drop images with geocode as 0,0
- Store the result to csv to use with dask dataframes

import re
import datetime

def format_usaf(code):
    return str(int(re.sub('\D', '', str(code))))

def year_moda_to_date(year, moda):
    try:
        return datetime.datetime(year,int(moda[:-2]), int(moda[-2:])).date()
    except:
        return None

#read and filter the images by accuracy for training?
df_images = pd.read_csv('../../../data/photo_metadata.csv', 
                        usecols = ['id', 'latitude', 'longitude','date_taken'])

df_images['latitude']=df_images['latitude'].round().astype(latitude_dtype)
df_images['longitude']=df_images['longitude'].round().astype(longitude_dtype)
df_images['geo'] = list(zip(df_images.latitude, df_images.longitude))

#Filter out missing geocode values
#optimize how the rows are filtered to reduce memory usage
drop_indexes = df_images[(df_images.latitude==0)&(df_images.longitude==0)].index
df_images = df_images.drop(drop_indexes)
drop_indexes = None

df_images['date_taken'] = pd.to_datetime(df_images['date_taken'], errors = 'coerce').map(lambda t: t.date())
df_images['date_taken_year'] = df_images['date_taken'].map(lambda t: t.year)

df_images = df_images[df_images.date_taken_year.isin([2007,2008,2009,2010,2011,2012])]
#df_images = df_images[df_images.date_taken_year.isin([2007])]

df_images.drop(columns=['date_taken_year'], inplace=True)

df_images.info()

df_images.to_pickle("../../../data/images_2007_2012.pickle")

df_images = pd.read_pickle("../../../data/images_2007_2012.pickle")

#df_images.to_csv("../../../data/images_processed.csv", index_label=False)

#df_images = dd.read_csv("../../../data/images_processed.csv")

### Process stations metadata

- Round the geocode for accuracy level to 111km
- Remove the stations that doesn't have any image
- Remove the stations without USAF

station_df= pd.read_csv('../../../data/isd-history.csv',
                        usecols = ['USAF', 'LAT', 'LON'])  

station_df.fillna(0, inplace=True)

station_df['LAT']=station_df['LAT'].round().astype(latitude_dtype)
station_df['LON']=station_df['LON'].round().astype(longitude_dtype)
station_df['geo'] = list(zip(station_df.LAT, station_df.LON))

#filter out the stations that doesn't have any image taken near by
print("Shape before filter", station_df.shape)

drop_indexes = station_df[~station_df.geo.isin(df_images.geo)].index
station_df = station_df.drop(drop_indexes)
drop_indexes = None

print("Shape after filter", station_df.shape)

station_df = station_df[station_df['USAF']!='999999']

station_df['USAF'] = station_df.USAF.swifter.progress_bar(True).apply(format_usaf)
station_df['USAF'] = station_df['USAF'].astype(int)
station_df.head(3)

station_df.to_pickle("../../../data/stations.pickle")

station_df = pd.read_pickle("../../../data/stations.pickle")

#station_df.to_csv("../../../data/station_processed.csv", index_label=False)

#station_df = pd.read_csv("../../../data/station_processed.csv")

#station_df['USAF'] = station_df['USAF'].astype(int)

station_df.info()

### Process Weather Data

- Drop the rows without STN code
- Create date field from year, moda fields
- Merge Wether data with stations data
- Keep one weather measures per geo per day

weather_df = pd.read_pickle("../../../data/weather_2007_2012.pkl")


weather_df = weather_df.astype({'STN':int, 'YEAR':'int16', 'MODA':'int16',
                                'TEMP':float,'DEWP':float,'SLP':float,
                                'VISIB':float, 'WDSP':float, 'PRCP':float, 'SNDP':float})

weather_df.info()
#24635200
#21697189

weather_df.reset_index(drop=True)
weather_df.head()

weather_df = weather_df[weather_df['STN']!=999999]

#weather_df = weather_df[['STN', 'YEAR', 'MODA','TEMP','DEWP', 'SLP','VISIB', 'WDSP','PRCP', 'SNDP']]

#weather_df['STN']=weather_df.swifter.progress_bar(True).apply(lambda x: format_usaf(str(x.STN)), axis=1)
#weather_df['STN']=weather_df['STN'].apply(format_usaf)

weather_df = weather_df[weather_df.STN.isin(station_df.USAF)]#.head(100000)

weather_df['date'] = weather_df.swifter.progress_bar(True).apply(lambda x: year_moda_to_date(int(x.YEAR), str(int(x.MODA))), axis=1)
#weather_df['date'] = weather_df.apply(lambda x: year_moda_to_date(x.YEAR, str(x.MODA)), axis=1)
weather_df.head()

weather_df.drop(columns=['YEAR', 'MODA'], inplace=True)

weather_df.to_pickle("../../../data/weather_2007_2012_processed.pkl")

#weather_df.to_csv("../../../data/weather_2007_2012_processed.csv", index_label=False)

weather_df = pd.read_pickle("../../../data/weather_2007_2012_processed.pkl")

del(weather_df)

#weather_df = pd.read_csv("../../../data/weather_2007_2012_processed.csv")

weather_merged = pd.merge(station_df, weather_df,  how='inner', left_on=['USAF'], right_on = ['STN'])

weather_merged.to_csv("../../../data/weather_merged.csv", index=False)

weather_merged.to_pickle("../../../data/weather_merged.pkl")

weather_merged = pd.read_pickle("../../../data/weather_merged.pkl")

#weather_merged = dd.read_csv("../../../data/weather_station_2007_2012.csv/*.part", usecols=['USAF', 'LAT','LON','STN', 'TEMP', 'DEWP', 'SLP', 'VISIB', 'WDSP', 'PRCP', 'SNDP', 'date'])

del(station_df)
del(weather_df)
gc.collect()

weather_merged = weather_merged.groupby(by=['LAT','LON','date']).first().reset_index()

weather_merged.head()

#weather_merged['LAT']=weather_merged['LAT'].round().astype(latitude_dtype)
#weather_merged['LON']=weather_merged['LON'].round().astype(longitude_dtype)

weather_merged['date'] = pd.to_datetime(weather_merged['date'], errors = 'coerce').map(lambda t: t.date())

weather_merged.to_pickle("../../../data/weather_station_2007_2012.pickle")

weather_merged = pd.read_pickle("../../../data/weather_station_2007_2012.pickle")

weather_merged.info()

### Merge weather data to Images metadata

img_weather = pd.merge(df_images, weather_merged,  how='inner', left_on=['latitude', 'longitude','date_taken'], right_on = ['LAT','LON','date'])

img_weather.head()

img_weather.columns

img_weather = img_weather[['id', 'date', 'LAT', 'LON', 'STN', 'TEMP', 'DEWP', 'SLP', 'VISIB', 'WDSP', 'PRCP', 'SNDP']]

img_weather.to_csv("../../../data/img_weather_2007_2012.csv", index=False)

img_weather.to_pickle("../../../data/img_weather_2007_2012.pickle")

# Charts

!pip install geopandas

import geopandas as gpd
from shapely.geometry import Point
import matplotlib.pyplot as plt

import descartes

df_images.head()

df_images.count()

import datetime

def datecast(strdate):
    try:
        return datetime.datetime.strptime(strdate, "%Y-%m-%d %H:%M:%S").date()
    except ValueError:
        return None

def get_date_diff(d1):
    return (datetime.date.today()-d1).days
  

df_images['dt_date_taken'] = df_images['date_taken'].apply(datecast)

df_images['Gap between taken and upload'] = df_images['dt_date_taken'].apply(get_date_diff)


# Creating histogram
from matplotlib import pyplot as plt
fig, ax = plt.subplots(figsize =(10, 7))
ax.hist(df_images['Gap between taken and upload'],edgecolor='black', align="mid",bins=[2000,3000,4000,5000,6000,7000,8000,9000,10000])
 
# Adding extra features   
plt.xlabel("Age of image(in days)")
plt.ylabel("Number of Images")
plt.title('Distribution of age of images')
#plt.xlim(0, 5000)
# Show plot
plt.show()

# Creating histogram
from matplotlib import pyplot as plt
fig, ax = plt.subplots(figsize =(10, 7))
ax.hist(df_images['accuracy'],edgecolor='black', align="mid",)
 
# Adding extra features   
plt.xlabel("Accuracy level")
plt.ylabel("Number of Images")
plt.title('Distribution of accuracy level- World level is 1, Country is ~3, Region ~6, City ~11, Street ~16')
# Show plot
plt.show()

# Creating histogram
from matplotlib import pyplot as plt
fig, ax = plt.subplots(figsize =(10, 7))
ax.hist(df_images['views'], bins=[0,5,10,15,20,25,30,35,40,45,50,55,60,65,70,75,80,85,90,95,100])
 
# Adding extra features   
plt.xlabel("Views")
plt.ylabel("# of Records")
plt.title('Distribution of views of images')
plt.xlim(0, 200)
#plt.ylim(0, 1000000)
# Show plot
plt.show()

location=[Point(xy) for xy in zip(df_images["longitude"],df_images["latitude"])]
location[:3]

crs={'init':'epsg:4326'}

geo_df=gpd.GeoDataFrame(df_images,crs=crs,geometry=location)
geo_df.head()

world = gpd.read_file(gpd.datasets.get_path('naturalearth_lowres'))
geo_df.plot(ax=world.plot(figsize=(20, 16)),markersize=2,color="red",marker=".")
