==> weather/step-5-prep-infer.ipynb <==
import pandas as pd

df_weather = pd.read_pickle("../../../data/weather_station_2007_2012.pickle")

def get_temperature_class(temperature):
    if temperature== 9999.9:
        return 99
    elif temperature>75:
        return 0
    elif 75>=temperature>65:
        return 1
    elif 65>=temperature>50:
        return 2
    else:
        return 3

def get_wind_speed_class(speed):
    if speed==999.9:
        return 99
    elif speed>8:
        return 0
    elif 8>=speed>4:
        return 1
    else:
        return 2
    
def get_dewp_class(dewp):
    if dewp==9999.9:
        return 99
    elif dewp>60:
        return 0
    elif 60>=dewp>45:
        return 1
    else:
        return 2
    
def get_visib_class(visib):
    if visib==999.9:
        return 99
    elif visib>8:
        return 0
    elif 8>=visib>6:
        return 1
    else:
        return 2

df_weather['DEWP_class'] = df_weather['DEWP'].apply(get_dewp_class)
df_weather['TEMP_class'] = df_weather['TEMP'].apply(get_temperature_class)
df_weather['VISIB_class'] = df_weather['VISIB'].apply(get_visib_class)
df_weather['WDSP_class'] = df_weather['WDSP'].apply(get_wind_speed_class)

df_weather.head(3)

df_weather.to_pickle("../../../data/weather_class.pkl")
